add api types
Use
openapi-typescript to convert the BE’s api’s yaml file to TS types.

This saves a lot of work on the FE side: in the case where an API
contract breaks,it’s possible to just update the local types using "npm
run update-api-types; npm run lint;" and check where the discrepancies
are.