setup store config
The Store.ts describes the top-level page’s <PageData>. Any children routes and component’s can have access to it.

I am however not setting the store from the server but from the client:
Server loads the data, passes it to the client, and the client in turn updates the local storage.
This design choice might be worth revisiting in the future.