Pagination
I added some pagination logic to display "…" once elements are out of visible range.
The pagination components pulls the page-number and total count of pages, and computes a limited range of page numbers to display.
The range is subscribed to the page’s stores - I tried pulling from "$store" instead, but that solution doesn’t seem to work.