Switch ApiHelpers to using generics
Since it was going to be necessary to find a reliable way to type api
responses’ data, I added a generic to be passed into the call
functions.
Sadly this is as much a curse as a blessing: passing types from the
ApiTypes is very verbose. For example:
```
api.call<
ApiTypes.components["responses"]["UserResponse"]["content"]["application/json"]>
```
is less clearly less than ideal. Maybe I’ll add some type aliases in the
future.