fix navigation for signIN/signOUT
navbar: added the data-sveltekit-reload functionality to force refresh
when signing out and when navigating back to front page. For some
reason, logged-in user can’t navigate back to front page without
requiring a hard refresh. Maybe the hooks are the source of the problem?
signup_signin: removed to use:enhance so that the form submission will
trigger hard refresh and full reload of the page, landing into logged-in
state.